Permissions + FexSeeding + Esim Sidebar Menu

+++ database/migrations/2024_08_26_104601_drop_column_tags_from_how_to_threads.php
@@ -14,7 +14,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('how_to_threads', function (Blueprint $table) {
-            $table->dropColumn(['tags']);
+            if (Schema::hasColumn('how_to_threads', 'tags')) {
+                $table->dropColumn(['tags']);
+            }
         });
     }
